[merge] : back <- back-develop

+++ back/iljungitjung/src/main/java/com/iljungitjung/domain/schedule/service/ReservationServiceImpl.java
@@ -19,10 +19,7 @@ import org.springframework.stereotype.Service;
 import javax.servlet.http.HttpSession;
 import javax.transaction.Transactional;
 import java.text.SimpleDateFormat;
-import java.util.ArrayList;
-import java.util.Calendar;
-import java.util.Date;
-import java.util.List;
+import java.util.*;
 
 @Service
 @RequiredArgsConstructor
@@ -30,7 +27,6 @@ public class ReservationServiceImpl implements ReservationService{
 
     private final ScheduleRepository scheduleRepository;
     private final CategoryRepository categoryRepository;
-
     private final UserRepository userRepository;
     private final UserService userService;
     private final NotificationService notificationService;
@@ -39,26 +35,38 @@ public class ReservationServiceImpl implements ReservationService{
     @Transactional
     public ReservationIdResponseDto reservationRequest(ReservationRequestDto reservationRequestDto, HttpSession httpSession) {
 
-        User user = userService.findUserBySessionId(httpSession);
+        User userFrom = userService.findUserBySessionId(httpSession);
 
         User userTo= userRepository.findUserByNickname(reservationRequestDto.getUserToNickname()).orElseThrow(() -> {
             throw new NoExistUserException();
         });
 
+        Schedule schedule = saveSchedule(userFrom, userTo, reservationRequestDto);
+
+        notificationService.autoReservationMessage(schedule);
+
+        return new ReservationIdResponseDto(schedule.getId());
+    }
+    private Schedule saveSchedule(User userFrom, User userTo, ReservationRequestDto reservationRequestDto){
+
         Category category = categoryRepository.findByCategoryNameAndUser_Email(reservationRequestDto.getCategoryName(), userTo.getEmail()).orElseThrow(() -> {
             throw new NoExistCategoryException();
         });
 
-        SimpleDateFormat formatter = new SimpleDateFormat("yyyyMMddHHmm");
-        Date startDate;
+        Date startDate = makeDateFormat(reservationRequestDto.getDate()+reservationRequestDto.getStartTime());
+        Date endDate = calcEndDate(startDate, category);
+
+        Schedule schedule = reservationRequestDto.toEntity(startDate, endDate, category.getColor());
+        schedule.setScheduleRequestList(userFrom);
+        schedule.setScheduleResponseList(userTo);
+
+        schedule = scheduleRepository.save(schedule);
+        return schedule;
+    }
+
+    private Date calcEndDate(Date startDate, Category category){
         String date = category.getTime();
         Calendar cal = Calendar.getInstance();
-
-        try{
-            startDate = formatter.parse(reservationRequestDto.getDate()+reservationRequestDto.getStartTime());
-        }catch (Exception e){
-            throw new DateFormatErrorException();
-        }
 
         cal.setTime(startDate);
         cal.add(Calendar.MINUTE, Integer.parseInt(date.substring(2)));
@@ -66,15 +74,8 @@ public class ReservationServiceImpl implements ReservationService{
 
         Date endDate = cal.getTime();
 
-        Schedule schedule = reservationRequestDto.toEntity(startDate, endDate, category.getColor());
-        schedule.setScheduleRequestList(user);
-        schedule.setScheduleResponseList(userTo);
-
-        schedule = scheduleRepository.save(schedule);
-        notificationService.autoReservationMessage(schedule);
-        return new ReservationIdResponseDto(schedule.getId());
+        return endDate;
     }
-
     @Override
     @Transactional
     public ReservationIdResponseDto reservationManage(Long id, ReservationManageRequestDto reservationManageRequestDto, HttpSession httpSession) {
@@ -85,27 +86,33 @@ public class ReservationServiceImpl implements ReservationService{
             throw new NoExistScheduleDetailException();
         });
 
-        String cancelFrom = "";
+        UserManageReservation(user, schedule, reservationManageRequestDto);
 
-        if(checkSamePerson(user, schedule.getUserTo())){
-            if(reservationManageRequestDto.isAccept()){
-                schedule.accepted();
-            }else{
-                cancelFrom="제공자";
-                schedule.canceled(cancelFrom, reservationManageRequestDto.getReason());
-            }
-        }else if(checkSamePerson(user, schedule.getUserFrom())){
-            if(reservationManageRequestDto.isAccept()){
-                throw new NoGrantAcceptScheduleException();
-            }else{
-                cancelFrom="사용자";
-                schedule.canceled(cancelFrom, reservationManageRequestDto.getReason());
-            }
-        }else{
-            throw new NoGrantAccessScheduleException();
-        }
         notificationService.autoReservationMessage(schedule);
+
         return new ReservationIdResponseDto(schedule.getId());
+    }
+
+    private void UserManageReservation(User user, Schedule schedule, ReservationManageRequestDto reservationManageRequestDto){
+        if(checkSameUser(user, schedule.getUserTo())) providerManageReservation(schedule, reservationManageRequestDto);
+        else if(checkSameUser(user, schedule.getUserFrom())) applicantManageReservation(schedule, reservationManageRequestDto);
+    }
+    private boolean checkSameUser(User userFrom, User userTo){
+        return userFrom.getId()==userTo.getId();
+    }
+    private void providerManageReservation(Schedule schedule, ReservationManageRequestDto reservationManageRequestDto){
+        if(reservationManageRequestDto.isAccept()){
+            schedule.accepted();
+            return;
+        }
+        schedule.canceled("제공자", reservationManageRequestDto.getReason());
+    }
+
+    private void applicantManageReservation(Schedule schedule, ReservationManageRequestDto reservationManageRequestDto){
+        if(reservationManageRequestDto.isAccept()){
+            throw new NoGrantAcceptScheduleException();
+        }
+        schedule.canceled("사용자", reservationManageRequestDto.getReason());
     }
 
     @Override
@@ -118,47 +125,58 @@ public class ReservationServiceImpl implements ReservationService{
             throw new NoExistScheduleException();
         });
 
-        if(!checkSamePerson(user, schedule.getUserTo())) throw new NoGrantDeleteScheduleException();
+        checkUserSchedule(user, schedule);
 
-        schedule.deleted();
         notificationService.autoReservationMessage(schedule);
+
+        scheduleRepository.delete(schedule);
+    }
+
+    private void checkUserSchedule(User user, Schedule schedule){
+        if (schedule.isUserSchedule(user)) {
+            schedule.deleted();
+            return;
+        }
+        throw new NoGrantDeleteScheduleException();
     }
 
     @Override
     @Transactional
     public ReservationBlockResponseDto reservationBlock(ReservationBlockListRequestDto reservationBlockListRequestDto, HttpSession httpSession) {
-        Long count = 0L;
         User user = userService.findUserBySessionId(httpSession);
 
-        List<Schedule> scheduleList = scheduleRepository.findByUserTo_IdIs(user.getId());
+        updateBlockDays(user, reservationBlockListRequestDto);
 
-        for(Schedule schedule : scheduleList){
-
-            if(schedule.getType().equals(Type.BLOCK)){
-                scheduleRepository.delete(schedule);
-            }
-        }
-
-        user.updateBlockDays(reservationBlockListRequestDto.getDays());
+        return makeReservationBlockResponseDto(user, reservationBlockListRequestDto);
+    }
+    private ReservationBlockResponseDto makeReservationBlockResponseDto(User user, ReservationBlockListRequestDto reservationBlockListRequestDto){
+        Long blockReservationCount=0L;
 
         for(ReservationBlockDto reservationBlockDto : reservationBlockListRequestDto.getBlockList()){
-            SimpleDateFormat formatter = new SimpleDateFormat("yyyyMMddHHmm");
-            Date startDate;
-            Date endDate;
-            try{
-                startDate = formatter.parse(reservationBlockDto.getDate()+reservationBlockDto.getStartTime());
-                endDate = formatter.parse(reservationBlockDto.getDate()+reservationBlockDto.getEndTime());
-            }catch (Exception e){
-                throw new DateFormatErrorException();
-            }
-            Schedule schedule = reservationBlockDto.toEntity(startDate, endDate);
-            schedule.setScheduleResponseList(user);
-            scheduleRepository.save(schedule);
-
-            count++;
+            saveBlockReservation(user, reservationBlockDto);
+            blockReservationCount++;
         }
-        return new ReservationBlockResponseDto(count);
 
+        return new ReservationBlockResponseDto(blockReservationCount);
+    }
+    private void saveBlockReservation(User user, ReservationBlockDto reservationBlockDto){
+        Date startDateFormat = makeDateFormat(reservationBlockDto.getDate()+reservationBlockDto.getStartTime());
+        Date endDateFormat = makeDateFormat(reservationBlockDto.getDate()+reservationBlockDto.getEndTime());
+
+        Schedule schedule = reservationBlockDto.toEntity(startDateFormat, endDateFormat);
+        schedule.blocked();
+        schedule.setScheduleResponseList(user);
+        scheduleRepository.save(schedule);
+    }
+
+    private void updateBlockDays(User user, ReservationBlockListRequestDto reservationBlockListRequestDto){
+        List<Schedule> scheduleList = scheduleRepository.findByUserTo_IdIs(user.getId());
+
+        scheduleList.forEach(schedule -> {
+            if(schedule.getType().equals(Type.BLOCK)) scheduleRepository.delete(schedule);
+        });
+
+        user.updateBlockDays(reservationBlockListRequestDto.getDays());
     }
 
     @Override
@@ -166,42 +184,57 @@ public class ReservationServiceImpl implements ReservationService{
 
         User user = userService.findUserBySessionId(httpSession);
 
-        Date startDateFormat;
-        Date endDateFormat;
+        return makeReservationViewResponseDto(user, startDate, endDate);
+    }
 
-        SimpleDateFormat formatter = new SimpleDateFormat("yyyyMMddHHmm");
-        try{
-            startDateFormat = formatter.parse(startDate+"0000");
-            endDateFormat = formatter.parse(endDate+"2359");
-        }catch (Exception e){
-            throw new DateFormatErrorException();
-        }
+    private ReservationViewResponseDto makeReservationViewResponseDto(User user, String startDate, String endDate){
 
+        startDate += "0000";
+        endDate += "2359";
 
+        Date startDateFormat = makeDateFormat(startDate);
+        Date endDateFormat = makeDateFormat(endDate);
+
+        return divideScheduleByType(user, startDateFormat, endDateFormat);
+    }
+
+    private ReservationViewResponseDto divideScheduleByType(User user, Date startDateFormat, Date endDateFormat){
         List<Schedule> scheduleList = scheduleRepository.findByUserFrom_IdIs(user.getId());
 
         List<ReservationViewDto> reservationViewDtoList = new ArrayList<>();
         List<ReservationCancelViewDto> reservationCancelViewDtoList = new ArrayList<>();
 
-        for(Schedule schedule : scheduleList){
-            if(checkDate(schedule, startDateFormat, endDateFormat)) continue;
+        scheduleList.forEach(schedule -> {
+            if(checkDate(schedule, startDateFormat, endDateFormat)) return;
 
-            if(schedule.getType().equals(Type.BLOCK)) continue;
-            if(schedule.getType().equals(Type.DELETE)) continue;
+            if(schedule.getType().equals(Type.BLOCK)) return;
+            if(schedule.getType().equals(Type.DELETE)) return;
+
             if(schedule.getType().equals(Type.CANCEL)) reservationCancelViewDtoList.add(new ReservationCancelViewDto(schedule));
             if(schedule.getType().equals(Type.REQUEST) || schedule.getType().equals(Type.ACCEPT)) reservationViewDtoList.add(new ReservationViewDto(schedule));
+        });
 
+        return new ReservationViewResponseDto(reservationViewDtoList, reservationCancelViewDtoList);
+    }
+
+    private boolean checkDate(Schedule schedule, Date startDateFormat, Date endDateFormat){
+        return schedule.getStartDate().before(startDateFormat)
+                || schedule.getEndDate().before(startDateFormat)
+                || schedule.getStartDate().after(endDateFormat)
+                || schedule.getEndDate().after(endDateFormat);
+    }
+
+    private Date makeDateFormat(String date){
+        Date dateFormat;
+
+        SimpleDateFormat formatter = new SimpleDateFormat("yyyyMMddHHmm");
+
+        try{
+            dateFormat = formatter.parse(date);
+        }catch (Exception e){
+            throw new DateFormatErrorException();
         }
-        ReservationViewResponseDto responseDto = new ReservationViewResponseDto(reservationViewDtoList, reservationCancelViewDtoList);
 
-        return responseDto;
-    }
-    public boolean checkDate(Schedule schedule, Date startDateFormat, Date endDateFormat){
-        if(schedule.getStartDate().before(startDateFormat) || schedule.getEndDate().before(startDateFormat) || schedule.getStartDate().after(endDateFormat) || schedule.getEndDate().after(endDateFormat)) return true;
-        return false;
-    }
-    public boolean checkSamePerson(User user1, User user2){
-        if(user1.getId()==user2.getId()) return true;
-        return false;
+        return dateFormat;
     }
 }
